feat(auth): add DTO classes for authentication module
- Create DTOs for user registration, login, password change, and password reset
- Add class-validator decorators for input validation
- Introduce UserProfileDto for returning user information
