+++ backend/src/auth/dto/change-password.dto.ts
@@ -1,0 +1,10 @@
+import { IsString, MinLength } from 'class-validator'
+
+export class ChangePasswordDto {
+  @IsString()
+  currentPassword: string
+
+  @IsString()
+  @MinLength(6)
+  newPassword: string
+}
