+++ backend/src/auth/dto/forgot-password.dto.ts
@@ -1,0 +1,6 @@
+import { IsEmail } from 'class-validator'
+
+export class ForgotPasswordDto {
+  @IsEmail()
+  email: string
+}
