+++ backend/src/auth/dto/index.ts
@@ -1,24 +1,6 @@
-export class RegisterDto {
-  username: string
-  email: string
-  password: string
-}
-
-export class LoginDto {
-  email: string
-  password: string
-}
-
-export class ChangePasswordDto {
-  currentPassword: string
-  newPassword: string
-}
-
-export class ForgotPasswordDto {
-  email: string
-}
-
-export class ResetPasswordDto {
-  token: string
-  newPassword: string
-}
+export * from './register.dto'
+export * from './login.dto'
+export * from './change-password.dto'
+export * from './forgot-password.dto'
+export * from './reset-password.dto'
+export * from './user-profile.dto'
