+++ backend/src/auth/dto/login.dto.ts
@@ -1,0 +1,9 @@
+import { IsEmail, IsString } from 'class-validator'
+
+export class LoginDto {
+  @IsEmail()
+  email: string
+
+  @IsString()
+  password: string
+}
