+++ backend/src/auth/dto/reset-password.dto.ts
@@ -1,0 +1,10 @@
+import { IsString, MinLength } from 'class-validator'
+
+export class ResetPasswordDto {
+  @IsString()
+  token: string
+
+  @IsString()
+  @MinLength(6)
+  newPassword: string
+}
