+++ backend/src/auth/dto/user-profile.dto.ts
@@ -1,0 +1,7 @@
+import { UserType } from '../../users/entities/user.enum'
+
+export class UserProfileDto {
+  username: string
+  email: string
+  role: UserType
+}
